Refactor of Microsoft.Bot.Builder.Azure.Queues.
