+++ libraries/Microsoft.Bot.Builder.Azure.Queues/AzureQueueStorage.cs
@@ -15,14 +15,14 @@ namespace Microsoft.Bot.Builder.Azure.Queues
     /// <summary>
     /// Service used to add messages to an Azure.Storage.Queues.
     /// </summary>
-    public class QueuesStorage
+    public class AzureQueueStorage : QueueStorage
     {
-        private JsonSerializerSettings _jsonSettings;
+        private readonly JsonSerializerSettings _jsonSettings;
         private bool _createQueueIfNotExists = true;
         private readonly QueueClient _queueClient;
 
         /// <summary>
-        /// Initializes a new instance of the <see cref="QueuesStorage"/> class.
+        /// Initializes a new instance of the <see cref="AzureQueueStorage"/> class.
         /// </summary>
         /// <param name="queuesStorageConnectionString">Azure Storage connection string.</param>
         /// <param name="queueName">Name of the storage queue where entities will be queued.</param>
@@ -31,7 +31,7 @@ namespace Microsoft.Bot.Builder.Azure.Queues
         /// <para>jsonSerializer.TypeNameHandling = TypeNameHandling.None.</para>
         /// <para>jsonSerializer.NullValueHandling = NullValueHandling.Ignore.</para>
         /// </param>
-        public QueuesStorage(string queuesStorageConnectionString, string queueName, JsonSerializerSettings jsonSerializerSettings = null)
+        public AzureQueueStorage(string queuesStorageConnectionString, string queueName, JsonSerializerSettings jsonSerializerSettings = null)
         {
             if (string.IsNullOrEmpty(queuesStorageConnectionString))
             {
@@ -44,25 +44,24 @@ namespace Microsoft.Bot.Builder.Azure.Queues
             }
 
             _jsonSettings = jsonSerializerSettings ?? new JsonSerializerSettings
-                {
-                    TypeNameHandling = TypeNameHandling.None,
-                    NullValueHandling = NullValueHandling.Ignore
-                };
+            {
+                TypeNameHandling = TypeNameHandling.None,
+                NullValueHandling = NullValueHandling.Ignore
+            };
 
             _queueClient = new QueueClient(queuesStorageConnectionString, queueName);
         }
 
         /// <summary>
-        /// Queue and Activity, with option in the Activity.Value to Azure.Storage.Queues.
+        /// Queue an Activity, with option in the Activity.Value to Azure.Storage.Queues.
         /// </summary>
         /// <param name="activity">This is expected to be an <see cref="Activity"/> retrieved from a call to 
         /// activity.GetConversationReference().GetContinuationActivity().  This enables restarting the conversation
         /// using BotAdapter.ContinueConversationAsync.</param>
-        /// <param name="visibilityTimeout">Default value of 0. Cannot be larger than 7 days.</param>
         /// <param name="timeToLive">Specifies the time-to-live interval for the message.</param>
         /// <param name="cancellationToken">Cancellation token for the async operation.</param>
         /// <returns><see cref="SendReceipt"/> as a Json string, from the QueueClient SendMessageAsync operation.</returns>
-        public async Task<string> QueueActivityAsync(Activity activity, TimeSpan? visibilityTimeout = null, TimeSpan? timeToLive = null,  CancellationToken cancellationToken = default)
+        public override async Task<string> QueueActivityAsync(Activity activity, TimeSpan? timeToLive = null, CancellationToken cancellationToken = default)
         {
             if (_createQueueIfNotExists)
             {
@@ -73,7 +72,7 @@ namespace Microsoft.Bot.Builder.Azure.Queues
             }
 
             var message = Convert.ToBase64String(Encoding.UTF8.GetBytes(JsonConvert.SerializeObject(activity, _jsonSettings)));
-            var receipt = await _queueClient.SendMessageAsync(message, visibilityTimeout, timeToLive, cancellationToken).ConfigureAwait(false);
+            var receipt = await _queueClient.SendMessageAsync(message, null, timeToLive, cancellationToken).ConfigureAwait(false);
 
             return JsonConvert.SerializeObject(receipt.Value);
         }
